requirements.txt
Django>=4.2,<5.0
psycopg[binary]>=3.1
python-dotenv>=1.0
python-pptx>=0.6.21
gunicorn>=22.0
whitenoise>=6.6
dj-database-url>=2.2
